+++ pinboard/app/helpers/users_helper.rb
@@ -1,0 +1,2 @@
+module UsersHelper
+end
