+++ pinboard/public/templates/signUpAndLogin.html
@@ -1,4 +1,5 @@
-<form name = "newUser">
+{{newUser}}
+<form name = "newUserForm">
   <label>email:
     <input type="email" ng-model="newUser.email">
   </label>
